+++ plugins/unleash-common/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @globallogicuki/backstage-plugin-unleash-common
 
+## 0.3.1
+
+## 0.3.0
+
+## 0.2.1
+
+### Patch Changes
+
+- 5983650: Add tag-based filtering for feature flags with interactive UI and annotation defaults
+
 ## 0.2.0
 
 ## 0.1.5